requirements.txt
# Core Deep Learning and Scientific Computing
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
scipy==1.10.1
pandas==2.0.1

# Data Processing
h5py==3.9.0
jsonlines==3.1.0
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
opencv-python==4.7.0.72

# Privacy and Security
cryptography==41.0.1
differential-privacy==0.6.0

# Logging and Monitoring
wandb==0.15.3
tensorboardX==2.6

# Configuration and Utility
pyyaml==6.0
tqdm==4.65.0

# Distributed Training Support
torch-distributed==2.1.0

# Optional: GPU Acceleration
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96

# Testing
pytest==7.3.1
unittest2==1.1.0

# Code Quality
mypy==1.3.0
black==23.3.0
flake8==6.0.0

# System and Performance Monitoring
psutil==5.9.5
GPUtil==1.4.0